+++ test/models/user_test.rb
@@ -33,4 +33,24 @@ class UserTest < ActiveSupport::TestCase
     assert_includes User.search("owner"), users(:business)
     assert_includes User.search("company"), users(:business)
   end
+
+  test "needs to sign the hiring agreement when active, not on a legacy plan, and not yet signed" do
+    user = users(:subscribed_business)
+
+    # Already signed.
+    refute user.needs_to_sign_hiring_agreement?
+
+    # Not signed.
+    hiring_agreements_signatures(:one).destroy
+    assert user.needs_to_sign_hiring_agreement?
+
+    # Not signed but not active.
+    HiringAgreements::Term.sole.deactivate!
+    refute user.needs_to_sign_hiring_agreement?
+
+    # Active and not signed but on legacy plan.
+    HiringAgreements::Term.sole.activate!
+    update_subscription(:legacy)
+    refute user.needs_to_sign_hiring_agreement?
+  end
 end